requirementes.txt
backports.tempfile==1.0
python-barcode==0.15.1
PIL==10.1.0
python-escpos==2.2.0
matplotlib==3.8.0
numpy==1.26.1
subprocess.run==0.0.8
tkPDFViewer==0.1
pdfkit==1.0.0
pdf2image==1.16.3
win32printing==0.1.3








 